+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 4.1.0 /
+
+## DEPENDENCIES
+
+- **chai** has been upgraded from 3.5.0 to **4.0.2**
+
 # 4.0.0 / 2017-05-04
 
 ## BREAKING CHANGES